+++ omniisaacgymenvs/scripts/rlgames_demo.py
@@ -47,7 +47,7 @@ class RLGDemo(RLGTrainer):
         self.cfg.test = True
     
 
-@hydra.main(config_name="config", config_path="../cfg")
+@hydra.main(version_base=None, config_name="config", config_path="../cfg")
 def parse_hydra_configs(cfg: DictConfig):
 
     time_str = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")