+++ omniisaacgymenvs/scripts/rlgames_train.py
@@ -84,7 +84,7 @@ class RLGTrainer():
         })
 
 
-@hydra.main(config_name="config", config_path="../cfg")
+@hydra.main(version_base=None, config_name="config", config_path="../cfg")
 def parse_hydra_configs(cfg: DictConfig):
 
     time_str = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")